requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==44.0.1
distro==1.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
greenlet==3.0.3
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
msgpack==1.1.0
oauthlib==3.2.2
openai==1.10.0
passlib==1.7.4
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.10.1
pyparsing==3.1.4
PyPDF2==3.0.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
